+++ pkg/providers/mtlstruststore/data_akamai_mtlstruststore_ca_set_activities.go
@@ -174,13 +174,13 @@ func (d *caSetActivitiesDataSource) Read(ctx context.Context, req datasource.Rea
 
 	if !data.Name.IsNull() {
 		tflog.Debug(ctx, "'name' provided, attempting to find CA set ID")
-		setID, err := findCASetID(ctx, client, data.Name.ValueString())
+		caSet, err := findNotDeletedCASetByName(ctx, client, data.Name.ValueString())
 		if err != nil {
 			resp.Diagnostics.AddError("Read CA set activities failed", err.Error())
 			return
 		}
 
-		data.ID = types.StringValue(setID)
+		data.ID = types.StringValue(caSet.CASetID)
 	}
 
 	activities, err := data.getActivities(ctx, client)
@@ -256,62 +256,4 @@ func (m *caSetActivitiesDataSourceModel) setData(activities caSetActivitiesDataS
 	m.DeletedDate = activities.DeletedDate
 	m.DeletedBy = activities.DeletedBy
 	m.Activities = activities.Activities
-}
-
-func findCASetID(ctx context.Context, client mtlstruststore.MTLSTruststore, caSetName string) (string, error) {
-	caSets, err := client.ListCASets(ctx, mtlstruststore.ListCASetsRequest{
-		CASetNamePrefix: caSetName,
-	})
-	if err != nil {
-		return "", fmt.Errorf("could not find CA Set ID for the given CA Set Name '%s', API error: %w", caSetName, err)
-	}
-
-	var matchingSets []mtlstruststore.CASetResponse
-	for _, caSet := range caSets.CASets {
-		if caSet.CASetName == caSetName {
-			matchingSets = append(matchingSets, caSet)
-		}
-	}
-
-	switch len(matchingSets) {
-	case 0:
-		return "", fmt.Errorf("no CA set found with name '%s'", caSetName)
-	case 1:
-		return matchingSets[0].CASetID, nil
-	default:
-		return "", fmt.Errorf("multiple CA sets IDs found with name '%s': %v. Use the ID to fetch a specific CA set", caSetName, buildSetsMap(matchingSets))
-	}
-}
-
-func findNotDeletedCASet(ctx context.Context, client mtlstruststore.MTLSTruststore, caSetName string) (mtlstruststore.CASetResponse, error) {
-	caSets, err := client.ListCASets(ctx, mtlstruststore.ListCASetsRequest{
-		CASetNamePrefix: caSetName,
-	})
-	if err != nil {
-		return mtlstruststore.CASetResponse{}, fmt.Errorf("could not find CA Set ID for the given CA set name '%s', API error: %w", caSetName, err)
-	}
-
-	var matchingSets []mtlstruststore.CASetResponse
-	for _, caSet := range caSets.CASets {
-		if caSet.CASetName == caSetName && caSet.CASetStatus == "NOT_DELETED" {
-			matchingSets = append(matchingSets, caSet)
-		}
-	}
-
-	switch len(matchingSets) {
-	case 0:
-		return mtlstruststore.CASetResponse{}, fmt.Errorf("no CA set found with name '%s' and status 'NOT_DELETED'", caSetName)
-	case 1:
-		return matchingSets[0], nil
-	default:
-		return mtlstruststore.CASetResponse{}, fmt.Errorf("multiple CA sets IDs found with name '%s' and status 'NOT_DELETED': %v. Use the ID to fetch a specific CA set", caSetName, buildSetsMap(matchingSets))
-	}
-}
-
-func buildSetsMap(sets []mtlstruststore.CASetResponse) map[string]string {
-	m := make(map[string]string)
-	for _, set := range sets {
-		m[set.CASetID] = set.CASetStatus
-	}
-	return m
 }