+++ pkg/providers/mtlstruststore/data_akamai_mtlstruststore_ca_set_versions.go
@@ -220,12 +220,12 @@ func (d *caSetVersionsDataSource) Read(ctx context.Context, req datasource.ReadR
 
 	if !data.Name.IsNull() {
 		tflog.Debug(ctx, "'name' provided, attempting to find CA set ID")
-		setID, err := findCASetID(ctx, client, data.Name.ValueString())
+		caSet, err := findNotDeletedCASetByName(ctx, client, data.Name.ValueString())
 		if err != nil {
 			resp.Diagnostics.AddError("Read CA set versions failed", err.Error())
 			return
 		}
-		data.ID = types.StringValue(setID)
+		data.ID = types.StringValue(caSet.CASetID)
 	}
 
 	if data.IncludeCertificates.IsNull() {