+++ pkg/providers/mtlstruststore/testdata/TestDataCASets/activated_on_with_deleting_status.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_mtlstruststore_ca_sets" "test" {
+  activated_on    = "PRODUCTION"
+  ca_set_statuses = ["DELETING"]
+}